Create Evaluating with METEOR
Evaluating with METEOR
METEOR excels at evaluating some of the more semantic features in text. It works similar to ROUGE by comparing a model-generated output to a reference output. You've been provided these texts as generated and reference; it's over to you to evaluate the score.

The evaluate library has been loaded for you.

Instructions

Compute and print the METEOR score.